+++ sandbox/grammars.js
@@ -1,6 +1,8 @@
 const main = grammarSourceCode => {
   const grammarConsole = $("#grammarConsole")
   const codeConsole = $("#codeConsole")
+  const codeErrorsConsole = $("#codeErrorsConsole")
+  const grammarErrorsConsole = $("#grammarErrorsConsole")
 
   // Init vars
   if (localStorage.getItem("grammarConsole")) grammarConsole.val(localStorage.getItem("grammarConsole"))
@@ -17,11 +19,14 @@ const main = grammarSourceCode => {
     .register()
     .fromTextAreaWithAutocomplete(grammarConsole[0], { lineWrapping: true })
 
-  grammarInstance.on("keyup", () => {
+  const updateGrammar = () => {
     const grammarCode = grammarInstance.getValue()
     localStorage.setItem("grammarConsole", grammarCode)
     window.grammarProgram = new GrammarConstructor(grammarCode)
-  })
+    const errs = window.grammarProgram.getProgramErrors()
+    grammarErrorsConsole.html(errs.length ? new TreeNode(errs).toTable() : "0 errors")
+  }
+  grammarInstance.on("keyup", updateGrammar)
 
   let grammarConstructor
   let cachedGrammarCode
@@ -49,11 +54,18 @@ const main = grammarSourceCode => {
     .register()
     .fromTextAreaWithAutocomplete(codeConsole[0], { lineWrapping: true })
 
-  codeInstance.on("keyup", () => {
+  const updateCode = () => {
     const code = codeInstance.getValue()
     localStorage.setItem("codeConsole", code)
     window.program = new grammarConstructor(code)
-  })
+    const errs = window.program.getProgramErrors()
+    codeErrorsConsole.html(new TreeNode(errs).toTable())
+  }
+
+  codeInstance.on("keyup", updateCode)
+
+  updateCode()
+  updateGrammar()
 }
 
 $(document).ready(function() {